feat(mgmt): reorganize models
