+++ mgmt/src/models/internal/interfaces/interface.rs
