+++ mgmt/src/models/internal/routing/statics.rs
